Added tool to automatically generate C interfaces to Objective-C objects.  Not connected to the build - it's only intended to be used internally, not distributed.
Note: This copies the AGSParser class from -base.  Unfortunately, this does not appear to be built into a shared library anywhere.  Ideally, it should be somewhere where both this tool and autogsdoc can link to it.

Usage:

$ make_c_interface -header /Local/Library/Headers/Foundation/NSArray.h

This will dump NSArray.plist and NSMutableArray.plist, containing suggested mappings from the methods in the two classes to functions.

¢ make_c_interface -classes NSArray -generateHeaders YES

This will generate a CFArray.h and CFArray.m file from the contents of NSArray.plist.  

You can modify the plist between running these commands, for example to remove methods that should not be exposed by the C interface or to change the order of parameters.  The ArgMappings key in the dictionary for each method stores the index of each C function argument that is passed to the Objective-C method.  You can tweak this for C interfaces that want to have their parameters passed in the opposite order to the methods.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/corebase/trunk@29815 72102866-910b-0410-8b05-ffd578937521